feat: Implement error handling in bookmark deletion process, addressing scenarios where the filtering operation encounters a ValidationError or no bookmarks found for deletion.Both scenarios now return a 'No bookmark found error'

+++ bookmarks/views/bookmark_bulk_delete_view.py
@@ -1,3 +1,5 @@
+from django.core.exceptions import ValidationError
+
 from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
@@ -8,8 +10,8 @@ from bookmarks.utils import group_bookmarks
 
 
 class BookmarkBulkDeleteAPIView(APIView):
-    bookmark_serializer = BookmarkSerializer
-    shortcut_serializer = ShortcutSerializer
+    bookmark_serializer_class = BookmarkSerializer
+    shortcut_serializer_class = ShortcutSerializer
 
     def delete(self, request, *args, **kwargs):
         bookmark_ids = request.data.get('ids', [])
@@ -21,16 +23,23 @@ class BookmarkBulkDeleteAPIView(APIView):
         all_bookmarks = Bookmark.objects.all()
 
         # Filter bookmarks to be deleted
-        bookmarks_to_delete = all_bookmarks.filter(id__in=bookmark_ids)
+        try:
+            bookmarks_to_delete = all_bookmarks.filter(id__in=bookmark_ids)
+        except (ValidationError) as error:
+            return Response(data={"error": "No bookmark found"}, status=status.HTTP_400_BAD_REQUEST)
+
+        if not bookmarks_to_delete:
+            return Response(data={"error": "No bookmark found"}, status=status.HTTP_400_BAD_REQUEST)
+
         bookmarks_to_delete.delete()
 
         # Exclude the deleted bookmarks from the original queryset
         all_bookmarks = all_bookmarks.exclude(id__in=bookmarks_to_delete.values('id'))
-        serialized_bookmarks = self.bookmark_serializer(all_bookmarks, many=True).data
+        serialized_bookmarks = self.bookmark_serializer_class(all_bookmarks, many=True).data
         grouped_bookmarks = group_bookmarks(serialized_bookmarks)
 
         shortcuts = all_bookmarks.filter(is_shortcut=True)
-        serialized_shortcuts = self.shortcut_serializer(shortcuts, many=True).data
+        serialized_shortcuts = self.shortcut_serializer_class(shortcuts, many=True).data
 
         response_data = {
             "message": "Bookmark deleted successfully.",
